requirements.txt
viam-sdk
pyinstaller
git+https://github.com/viam-labs/1liner
smbus2